Changes to make appveyor script usable for developers. - Autodetect Visual Studio version, edition or buildtools. Default to predefined buildtools as before, prefer 2017 to 2019 and Enterprise to Professional to Community. Explicit parameter to make VC 2019 to use VC 2017 (14.1) compiler. - Autodetect path of 64bit binaries so no hardcoded path anymore. - Two new parameters to skip build and run tests and skip installer generation. - Install_buildenv first checks if path exists before checking if platform has 7zip. - Added alternative using powershell if 7zip not found. This is all complemented with a simple build.cmd that calls these two scripts which will be placed on the wiki.

+++ build/windows/install_buildenv.bat
@@ -10,30 +10,46 @@ SET WINLIB_NAME=%2
 REM The local directory to cache build environments in.
 SET WINLIBS_PATH=%3
 
-REM Install build env base dir
-IF NOT EXIST "%WINLIBS_PATH%" MKDIR "%WINLIBS_PATH%"
-
-where /q 7za
-IF ERRORLEVEL 1 (
-  where /q 7z
-  if ERRORLEVEL 1 (
-    echo Could not find 7z or 7za on the path.
-    exit /b
-  ) else (
-    set ZIP=7z
-  )
-) else (
-  set ZIP=7za
-)
-
 IF NOT EXIST "%WINLIBS_PATH%\%WINLIB_NAME%" (
+
+  REM Install build env base dir
+  IF NOT EXIST "%WINLIBS_PATH%" MKDIR "%WINLIBS_PATH%"
+
+  where /q 7za
+  IF ERRORLEVEL 1 (
+    where /q 7z
+    if ERRORLEVEL 1 (
+      where /q powershell
+      if ERRORLEVEL 1 (
+         echo Could not find 7z or 7za on the path, and powershell is not available. 
+         echo Either install http://7-zip.org
+         echo or download %BASEURL%/%WINLIB_NAME%.zip
+         echo yourself and unzip it at %WINLIBS_PATH%
+         ENDLOCAL && EXIT /b 1
+      ) else (
+        set POWERSHELL=powershell
+      )
+    ) else (
+      set ZIP=7z
+    )
+  ) else (
+    set ZIP=7za
+  )
+
   echo Installing winlib %WINLIB_NAME%.
   CD "%WINLIBS_PATH%"
   echo ..Downloading %WINLIB_NAME%.zip
   curl -fsS -L -o %WINLIB_NAME%.zip %BASEURL%/%WINLIB_NAME%.zip
+  IF ERRORLEVEL 1 ENDLOCAL && EXIT /b 1
   REM TODO(XXX) check fingerprint on zip file.
   echo ..unzipping %WINLIB_NAME%.zip
-  %ZIP% x %WINLIB_NAME%.zip
+  IF "%POWERSHELL%" == "powershell" (
+    powershell -Command "& {Expand-Archive \"%WINLIBS_PATH%/%WINLIB_NAME%.zip\" \"%WINLIBS_PATH%\\\"}"
+    IF ERRORLEVEL 1 ENDLOCAL && EXIT /b 1
+  ) ELSE (
+    %ZIP% x %WINLIB_NAME%.zip
+    IF ERRORLEVEL 1 ENDLOCAL && EXIT /b 1
+  )
   DEL %WINLIB_NAME%.zip
 ) else (
   echo %WINLIB_NAME% already exists at %WINLIBS_PATH%
